Refactor y documentar la conversión de NFA a DFA, eliminando archivos obsoletos y mejorando la claridad del código.
