+++ Compiler/src/lexgen/dfa/mod.rs
@@ -1,3 +1,2 @@
 pub mod dfa;
-pub mod state;
 pub mod utils;
