--- Compiler/src/lexgen/dfa/state.rs
@@ -1,1 +1,0 @@
-// Definición de estados del DFA
