+++ Compiler/src/lexgen/nfa/builder/mod.rs
@@ -1,0 +1,5 @@
+// ===============================
+// mod.rs: Módulo de builder para NFA
+// ===============================
+
+pub mod builder;
