+++ Compiler/src/lexgen/nfa/mod.rs
@@ -1,3 +1,4 @@
+pub mod builder;
 pub mod join_nfa;
 pub mod nfa;
 pub mod state;
