+++ Compiler/src/lexgen/regex_parser/node/un_op.rs
@@ -23,4 +23,3 @@ impl RegexUnOp {
         }
     }
 }
-
