+++ Compiler/src/lexgen/regex_parser/regex_parser.rs
@@ -1,3 +1,8 @@
+// ===============================
+// regex_parser.rs: Parser de Expresiones Regulares a AST
+// ===============================
+// Convierte una cadena regex en un AST modularizado y documentado.
+
 use crate::regex_parser::node::ast_node_impl::{AstNode, AstNodeImpl, AstNodeKind};
 use crate::regex_parser::node::bin_op::RegexBinOp;
 use crate::regex_parser::node::group::RegexGroup;
@@ -7,6 +12,7 @@ use crate::regex_parser::node::regex_escape::RegexEscape;
 use crate::regex_parser::node::un_op::RegexUnOp;
 
 /// Parsea una expresión regular y retorna el AST correspondiente.
+/// Soporta alternancia, concatenación, agrupación, clases, escapes y operadores unarios.
 pub fn parse_regex(input: &str) -> Option<AstNodeImpl> {
     let input = input.trim();
     if input.is_empty() {
@@ -25,8 +31,8 @@ pub fn parse_regex(input: &str) -> Option<AstNodeImpl> {
         let c = input.chars().next().unwrap();
         return Some(LiteralNode::new(c).to_ast());
     }
-    // Alternancia (|)
-    if let Some(idx) = find_top_level(input, '|') {
+    // Alternancia (|) de nivel superior
+    if let Some(idx) = helpers::find_top_level(input, '|') {
         let left = &input[..idx];
         let right = &input[idx + 1..];
         let left_ast = parse_regex(left)?;
@@ -39,8 +45,8 @@ pub fn parse_regex(input: &str) -> Option<AstNodeImpl> {
             },
         });
     }
-    // Concatenación implícita
-    if let Some(idx) = find_concat_point(input) {
+    // Concatenación implícita de nivel superior
+    if let Some(idx) = helpers::find_concat_point(input) {
         let left = &input[..idx];
         let right = &input[idx..];
         let left_ast = parse_regex(left)?;
@@ -53,8 +59,8 @@ pub fn parse_regex(input: &str) -> Option<AstNodeImpl> {
             },
         });
     }
-    // Paréntesis (agrupación)
-    if input.starts_with('(') && input.ends_with(')') && is_balanced_parens(input) {
+    // Agrupación con paréntesis
+    if input.starts_with('(') && input.ends_with(')') && helpers::is_balanced_parens(input) {
         let inner = &input[1..input.len() - 1];
         let expr = parse_regex(inner)?;
         return Some(AstNodeImpl {
@@ -62,7 +68,7 @@ pub fn parse_regex(input: &str) -> Option<AstNodeImpl> {
         });
     }
     // Operadores unarios: *, +, ?
-    if let Some((body, op)) = parse_unary_suffix(input) {
+    if let Some((body, op)) = helpers::parse_unary_suffix(input) {
         let expr = parse_regex(body)?;
         let op = match op {
             '*' => RegexUnOp::Star,
@@ -80,7 +86,7 @@ pub fn parse_regex(input: &str) -> Option<AstNodeImpl> {
     // Clase de caracteres simple: [abc], [a-z]
     if input.starts_with('[') && input.ends_with(']') {
         let inner = &input[1..input.len() - 1];
-        let class = parse_class(inner)?;
+        let class = helpers::parse_class(inner)?;
         return Some(AstNodeImpl {
             kind: AstNodeKind::Class(class),
         });
@@ -96,73 +102,80 @@ pub fn parse_regex(input: &str) -> Option<AstNodeImpl> {
     None
 }
 
-/// Busca el índice de un operador binario de nivel superior (no anidado en paréntesis).
-fn find_top_level(input: &str, op: char) -> Option<usize> {
-    let mut depth = 0;
-    for (i, c) in input.chars().enumerate() {
-        match c {
-            '(' => depth += 1,
-            ')' => depth -= 1,
-            _ if c == op && depth == 0 => return Some(i),
-            _ => {}
-        }
-    }
-    None
-}
+// ===============================
+// helpers: Submódulo privado para utilidades de parsing
+// ===============================
+mod helpers {
+    use super::*;
 
-/// Busca el punto de concatenación implícita de nivel superior.
-fn find_concat_point(input: &str) -> Option<usize> {
-    let mut depth = 0;
-    let chars: Vec<char> = input.chars().collect();
-    for i in 1..chars.len() {
-        match chars[i] {
-            '(' => depth += 1,
-            ')' => depth -= 1,
-            _ => {}
-        }
-        if depth == 0 {
-            // No separar en medio de un grupo
-            return Some(i);
-        }
-    }
-    None
-}
-
-/// Verifica si los paréntesis están balanceados.
-fn is_balanced_parens(input: &str) -> bool {
-    let mut depth = 0;
-    for (i, c) in input.chars().enumerate() {
-        match c {
-            '(' => depth += 1,
-            ')' => {
-                depth -= 1;
-                if depth == 0 && i != input.len() - 1 {
-                    return false;
-                }
+    /// Busca el índice de un operador binario de nivel superior (no anidado en paréntesis).
+    pub fn find_top_level(input: &str, op: char) -> Option<usize> {
+        let mut depth = 0;
+        for (i, c) in input.chars().enumerate() {
+            match c {
+                '(' => depth += 1,
+                ')' => depth -= 1,
+                _ if c == op && depth == 0 => return Some(i),
+                _ => {}
             }
-            _ => {}
         }
+        None
     }
-    depth == 0
-}
 
-/// Parsea un operador unario al final de la expresión.
-fn parse_unary_suffix(input: &str) -> Option<(&str, char)> {
-    if input.len() < 2 {
-        return None;
-    }
-    let last = input.chars().last().unwrap();
-    if last == '*' || last == '+' || last == '?' {
-        let body = &input[..input.len() - 1];
-        if !body.is_empty() {
-            return Some((body, last));
+    /// Busca el punto de concatenación implícita de nivel superior.
+    pub fn find_concat_point(input: &str) -> Option<usize> {
+        let mut depth = 0;
+        let chars: Vec<char> = input.chars().collect();
+        for i in 1..chars.len() {
+            match chars[i] {
+                '(' => depth += 1,
+                ')' => depth -= 1,
+                _ => {}
+            }
+            if depth == 0 {
+                // No separar en medio de un grupo
+                return Some(i);
+            }
         }
+        None
     }
-    None
-}
 
-/// Parsea una clase de caracteres simple.
-fn parse_class(input: &str) -> Option<RegexClass> {
-    let chars: Vec<RegexChar> = input.chars().map(RegexChar::Literal).collect();
-    Some(RegexClass::Set(chars))
+    /// Verifica si los paréntesis están balanceados.
+    pub fn is_balanced_parens(input: &str) -> bool {
+        let mut depth = 0;
+        for (i, c) in input.chars().enumerate() {
+            match c {
+                '(' => depth += 1,
+                ')' => {
+                    depth -= 1;
+                    if depth == 0 && i != input.len() - 1 {
+                        return false;
+                    }
+                }
+                _ => {}
+            }
+        }
+        depth == 0
+    }
+
+    /// Parsea un operador unario al final de la expresión.
+    pub fn parse_unary_suffix(input: &str) -> Option<(&str, char)> {
+        if input.len() < 2 {
+            return None;
+        }
+        let last = input.chars().last().unwrap();
+        if last == '*' || last == '+' || last == '?' {
+            let body = &input[..input.len() - 1];
+            if !body.is_empty() {
+                return Some((body, last));
+            }
+        }
+        None
+    }
+
+    /// Parsea una clase de caracteres simple (solo set, no rangos ni negaciones).
+    pub fn parse_class(input: &str) -> Option<RegexClass> {
+        let chars: Vec<RegexChar> = input.chars().map(RegexChar::Literal).collect();
+        Some(RegexClass::Set(chars))
+    }
 }
